+++ packages/ricos-editor/web/src/RicosEditorSwitcher.tsx
@@ -1,9 +1,9 @@
-import React, { forwardRef } from 'react';
+import React, { forwardRef, Suspense } from 'react';
 import type { RicosEditorProps } from 'ricos-common';
 import type { RicosEditorRef } from './RicosEditorRef';
 import RicosEditorWithRef from './RicosEditor';
 import type { RicosEditor } from './RicosEditor';
-import type FullRicosEditorTiptap from './RicosEditorTiptap/FullRicosEditorTiptap';
+import type { FullRicosEditorTiptap } from './RicosEditorTiptap/FullRicosEditorTiptap';
 
 const FullRicosEditorTiptapLazy = React.lazy(
   /* webpackChunkName: "FullRicosEditorTiptap" */
@@ -14,10 +14,12 @@ const RicosEditorSwitcher = React.forwardRef<RicosEditorRef, RicosEditorProps>((
   const useTiptap = !!props.experiments?.tiptapEditor?.enabled;
   if (useTiptap) {
     return (
-      <FullRicosEditorTiptapLazy
-        {...props}
-        ref={ref as React.ForwardedRef<FullRicosEditorTiptap>}
-      />
+      <Suspense fallback={<div />}>
+        <FullRicosEditorTiptapLazy
+          {...props}
+          ref={ref as React.ForwardedRef<FullRicosEditorTiptap>}
+        />
+      </Suspense>
     );
   } else {
     return <RicosEditorWithRef {...props} ref={ref as React.ForwardedRef<RicosEditor>} />;